fix(backoffice-v2): fixed wrong file location for download file svg
